# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sistema de Calificaciones - API",
    "version": "0.2.0",
    "description": "API REST del Sistema de Calificaciones Escolar. Multi-tenencia por escuela. Autenticación mediante cookie de sesión firmada (HMAC-SHA256) o Bearer token (para automatizaciones)."
  },
  "servers": [
    { "url": "/", "description": "Servidor actual" }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Cookie de sesión firmada con HMAC-SHA256."
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token para automatizaciones (AGENT_SECRET_TOKEN). Solo para endpoints específicos."
      }
    },
...
  },
  "security": [{ "sessionCookie": [] }],
  "paths": {
...
    "/api/boleta": {
      "get": {
        "summary": "Generar boleta de calificaciones",
        "parameters": [
          { "name": "estudianteId", "in": "query", "schema": { "type": "string" } },
          { "name": "gradoId", "in": "query", "schema": { "type": "string" } },
          { "name": "trimestre", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Boleta generada" } }
      }
    },
...
    "/api/agent/monitor": {
      "get": { "summary": "Obtener alertas del agente", "responses": { "200": { "description": "Alertas activas" } } },
      "post": {
        "summary": "Ejecutar análisis de riesgo",
        "security": [{ "sessionCookie": [] }, { "bearerToken": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "año": { "type": "integer" },
                  "gradoId": { "type": "string" },
                  "guardarAlertas": { "type": "boolean" },
                  "escuelaId": { "type": "string", "description": "Requerido con Bearer token" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Análisis completado" } }
      },
      "delete": { "summary": "Borrar alertas/logs del agente", "responses": { "200": { "description": "Alertas borradas" } } }
    },
...
  }
}